183: Reorganize core
